test(runner): separate request-hook firewall auth tests
